Pass config to web; extract json serialization from the model
